lesson 7: add role model

+++ api/handlers.py
@@ -12,6 +12,7 @@ from api.actions.user import _create_new_user
 from api.actions.user import _delete_user
 from api.actions.user import _get_user_by_id
 from api.actions.user import _update_user
+from api.actions.user import check_user_permissions
 from api.models import DeleteUserResponse
 from api.models import ShowUser
 from api.models import UpdatedUserResponse
@@ -42,6 +43,16 @@ async def delete_user(
     session: AsyncSession = Depends(get_db),
     current_user: User = Depends(get_current_user_from_token),
 ) -> DeleteUserResponse:
+    user_for_deletion = await _get_user_by_id(user_id, session)
+    if user_for_deletion is None:
+        raise HTTPException(
+            status_code=404, detail=f"User with id {user_id} not found."
+        )
+    if not check_user_permissions(
+        target_user=user_for_deletion,
+        current_user=current_user,
+    ):
+        raise HTTPException(status_code=403, detail="Forbidden.")
     deleted_user_id = await _delete_user(user_id, session)
     if deleted_user_id is None:
         raise HTTPException(
@@ -79,12 +90,18 @@ async def update_user_by_id(
             status_code=422,
             detail="At least one parameter for user update info should be provided",
         )
-    user = await _get_user_by_id(user_id, session)
-    if user is None:
+    user_for_update = await _get_user_by_id(user_id, session)
+    if user_for_update is None:
         raise HTTPException(
             status_code=404,
             detail=f"User with id {user_id} not found.",
         )
+    if user_id != current_user.user_id:
+        if check_user_permissions(
+            target_user=user_for_update,
+            current_user=current_user,
+        ):
+            raise HTTPException(status_code=403, detail="Forbidden.")
     try:
         updated_user_id = await _update_user(
             updated_user_params=updated_user_params,
